initial commit of the bulk run tool
This is a *very* early prototype.  Currently it can submit a single job
listed on its command line, or accept any number of jobs as command-line
formatted command-per-line files or standard input through the -c
argument, which may be repeated.  To try the ATS tests, run them with
`flux capacitor -c <filename>`, but make sure your file-descriptor limit
is high first, very very high.

The tool is currently not using the sched module, because I was running
into issues with stalls and jobs getting ignored.  This is probably
because of a resource specification issue, so I'll revisit that when the
sched module can interpret the hwloc resource output, or has some other
way to infer its environment.  For now, the tool itself implements a
basic scheduler with no concept of ordering fairness, it just goes for
maximum utilization without oversubscription, and only accepts a maximum
of 100 pending jobs at any given time.  Whenever it reaches 100 pending
jobs, it enters the reactor repeatedly in 0.5 second increments until
that number drops below 100, then adds more.  Wash, rinse, repeat in
that fashion until all specified commands are submitted.  Once that has
happened, it falls into a final reactor start call with a timeout of 10
minutes.  That's there as a fallback because something in flux eats the
<C-c> key command, so canceling one of these is problematic at the
moment.  I'm not sure about the best way to fix it but it's on the list.

The MPI test does not seem to be working for me at the moment, but that
appears to be an issue with my build environment rather than the tool
itself, will debug this as it goes forward.